cannot delete a category that has other categories as children

+++ app/Http/Controllers/Admin/CategoryController.php
@@ -58,8 +58,14 @@ class CategoryController extends Controller
     {
         if ($category->articles()->exists()) {
             return redirect(route('admin.categories'))
-                ->with('type', 'into')
+                ->with('type', 'info')
                 ->with('status', 'Category was not deleted. There are articles associated to it.');
+        }
+
+        if($category->children()->exists()) {
+            return redirect(route('admin.categories'))
+                ->with('type', 'info')
+                ->with('status', 'Category was not deleted. There are categories associated to it.');
         }
 
         $category->delete();
